+++ aftiktuna-rust/aftiktuna-three-d/src/game/ui.rs
@@ -1,11 +1,12 @@
 use crate::asset::{Assets, LazilyLoadedModels};
 use aftiktuna::command_suggestion::Suggestion;
-use aftiktuna::view::Frame;
+use aftiktuna::view::{Frame, FullStatus};
 use three_d::egui;
 
 #[derive(Default)]
 pub struct UiResult {
     pub triggered_input: bool,
+    pub closed_status_window: bool,
     pub clicked_text_box: bool,
     pub clicked_suggestion: Option<Suggestion>,
 }
@@ -16,6 +17,15 @@ pub fn update_ui(
     state: &mut super::State,
     assets: &mut Assets,
 ) -> UiResult {
+    gui.context().style_mut(|style| {
+        style.spacing.menu_margin = egui::Margin::ZERO;
+        style.spacing.item_spacing = egui::Vec2::ZERO;
+        style.visuals.menu_rounding = egui::Rounding::ZERO;
+        style.visuals.popup_shadow = egui::Shadow::NONE;
+        style.visuals.window_fill = egui::Color32::TRANSPARENT;
+        style.visuals.window_stroke = egui::Stroke::NONE;
+    });
+
     let mut ui_result = UiResult::default();
     gui.update(
         &mut frame_input.events,
@@ -23,19 +33,91 @@ pub fn update_ui(
         frame_input.viewport,
         frame_input.device_pixel_ratio,
         |egui_context| {
+            egui_context.style_mut(|style| {
+                style.visuals.override_text_color = None;
+                style.visuals.widgets.noninteractive.bg_stroke =
+                    egui::Stroke::new(1., egui::Color32::from_gray(60));
+            });
+
             ui_result.triggered_input = input_panel(
                 &mut state.input_text,
-                state.game.ready_to_take_input(),
+                state.game.ready_to_take_input() && state.displayed_status.is_none(),
                 std::mem::take(&mut state.request_input_focus),
                 egui_context,
             );
-            ui_result.clicked_text_box = text_box_panel(&state.text_box_text, egui_context);
 
-            ui_result.clicked_suggestion =
-                show_tooltip_and_menu(state, &mut assets.models, egui_context);
+            egui_context.style_mut(|style| {
+                style.visuals.override_text_color = Some(egui::Color32::WHITE);
+                style.visuals.widgets.noninteractive.bg_stroke =
+                    egui::Stroke::new(1., egui::Color32::WHITE);
+            });
+
+            if !state.text_box_text.is_empty() {
+                ui_result.clicked_text_box = text_box_panel(&state.text_box_text, egui_context);
+            }
+
+            if let Some(status) = &state.displayed_status {
+                ui_result.closed_status_window = !show_status_screen(status, egui_context);
+            } else {
+                ui_result.clicked_suggestion =
+                    show_tooltip_and_menu(state, &mut assets.models, egui_context);
+            }
         },
     );
     ui_result
+}
+
+const STATUS_DISPLAY_OUTER_MARGIN: egui::Vec2 = egui::vec2(180., 50.);
+const STATUS_WINDOW_END_COMPENSATION: egui::Vec2 = egui::vec2(24., 43.);
+const STATUS_DISPLAY_COLOR: egui::Color32 = egui::Color32::from_rgba_premultiplied(
+    (0.25 * 0.95 * 255.) as u8,
+    (0.2 * 0.95 * 255.) as u8,
+    (0.3 * 0.95 * 255.) as u8,
+    (0.95 * 255.) as u8,
+);
+
+fn show_status_screen(status: &FullStatus, egui_context: &egui::Context) -> bool {
+    let mut is_open = true;
+    egui::Window::new(egui::RichText::new("Crew Status").font(egui::FontId::monospace(15.0)))
+        .open(&mut is_open)
+        .collapsible(false)
+        .fixed_rect(
+            egui::Rect::from_min_max(
+                egui::Pos2::ZERO,
+                egui::pos2(
+                    crate::WINDOW_WIDTH_F,
+                    crate::WINDOW_HEIGHT_F - INPUT_PANEL_HEIGHT,
+                ) - STATUS_WINDOW_END_COMPENSATION,
+            )
+            .shrink2(STATUS_DISPLAY_OUTER_MARGIN),
+        )
+        .frame(egui::Frame {
+            inner_margin: egui::Margin::symmetric(12., 6.),
+            fill: STATUS_DISPLAY_COLOR,
+            stroke: egui::Stroke::new(1., egui::Color32::WHITE),
+            ..Default::default()
+        })
+        .vscroll(true)
+        .show(egui_context, |ui| {
+            for line in &status.ship {
+                ui.label(
+                    egui::RichText::new(line)
+                        .font(TEXT_BOX_FONT)
+                        .line_height(Some(15.)),
+                );
+            }
+            for character_text in &status.crew {
+                ui.separator();
+                for line in character_text {
+                    ui.label(
+                        egui::RichText::new(line)
+                            .font(TEXT_BOX_FONT)
+                            .line_height(Some(13.)),
+                    );
+                }
+            }
+        });
+    is_open
 }
 
 fn show_tooltip_and_menu(
@@ -56,9 +138,7 @@ fn show_tooltip_and_menu(
                         .inner_margin(egui::Margin::symmetric(TEXT_BOX_MARGIN, 1.))
                         .begin(ui);
 
-                    prepared
-                        .content_ui
-                        .label(egui::RichText::new(suggestion.text()).color(egui::Color32::WHITE));
+                    prepared.content_ui.label(suggestion.text());
 
                     let response = prepared.allocate_space(ui).interact(egui::Sense::click());
                     if response.clicked() {
@@ -99,7 +179,7 @@ fn show_tooltip_and_menu(
                             .fill(TEXT_BOX_COLOR)
                             .inner_margin(egui::Margin::symmetric(TEXT_BOX_MARGIN, 1.))
                             .show(ui, |ui| {
-                                ui.label(egui::RichText::new(line).color(egui::Color32::WHITE));
+                                ui.label(line);
                             });
                     }
                 },
@@ -175,8 +255,7 @@ fn text_box_panel<S: Into<String>>(
                         ui.label(
                             egui::RichText::new(line)
                                 .font(TEXT_BOX_FONT)
-                                .line_height(Some(13.))
-                                .color(egui::Color32::WHITE),
+                                .line_height(Some(13.)),
                         );
                     }
                 });
@@ -230,14 +309,6 @@ fn show_hovering<T>(
     pos: three_d::Vec2,
     content: impl Fn(&mut egui::Ui) -> T,
 ) -> T {
-    egui_context.style_mut(|style| {
-        style.spacing.menu_margin = egui::Margin::ZERO;
-        style.spacing.item_spacing = egui::Vec2::ZERO;
-        style.visuals.menu_rounding = egui::Rounding::ZERO;
-        style.visuals.popup_shadow = egui::Shadow::NONE;
-        style.visuals.window_fill = egui::Color32::TRANSPARENT;
-        style.visuals.window_stroke = egui::Stroke::NONE;
-    });
     egui::show_tooltip_at(
         egui_context,
         egui::LayerId::background(),